requirements.txt
streamlit
opencv-python
pytesseract
easyocr
openai
reportlab
Pillow
